Preliminary sql/soap client
by design, the two applications are separated.
The goal would be to provide (in addition) a all-in-one solution with
one thread for the SOAP server and one thread for the SQL feeder.

There should also be a basic client to turn on/off the PS and to
enable/disable db logging.