Address some of the problems spotted with the dev
These are not really bugs (except for settings, still pending) but non
practical issues, like the impossibility to know what is going on via
the display and the impossibility to get the error messages from the
hardware.

Also introduced the error logging to file in the client, as the console
is not available in that case.